Refactored and moved post fragments into one folder and renamed activity bar title.
