+++ app/src/main/java/com/example/seniorproject/viewModels/NewPostFragmentViewModel.kt
@@ -1,26 +1,13 @@
 package com.example.seniorproject.viewModels
 
-import android.content.Context
 import android.net.Uri
 import android.util.Log
 import android.view.View
 import android.widget.AdapterView
-import android.widget.Toast
-import androidx.core.net.toUri
 import androidx.lifecycle.ViewModel
-import com.example.seniorproject.MainForum.Fragments.FragmentNewPost
-import com.example.seniorproject.MainForum.Posts.UpdatePost
 //import com.example.seniorproject.Utils.startMainForum
 import com.example.seniorproject.Utils.PostListener
-import com.example.seniorproject.data.models.Post
 import com.example.seniorproject.data.repositories.PostRepository
-import com.google.firebase.database.DataSnapshot
-import com.google.firebase.storage.FirebaseStorage
-import com.google.firebase.storage.UploadTask
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.GlobalScope
-import kotlinx.coroutines.delay
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class NewPostFragmentViewModel @Inject constructor(private val repository: PostRepository) : ViewModel() {
