+++ app/src/main/res/layout/fragment_image__post.xml
@@ -14,7 +14,7 @@
             android:layout_width="match_parent"
             android:background="@color/white"
             android:layout_height="wrap_content"
-            tools:context=".MainForum.Fragments.FragmentNewPost">
+            tools:context=".MainForum.NewPost.FragmentNewPost">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/cl"
