+++ app/src/main/res/layout/fragment_new_post.xml
@@ -21,7 +21,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:background="@color/BackgroundColor"
-            tools:context=".MainForum.Fragments.FragmentNewPost">
+            tools:context=".MainForum.NewPost.FragmentNewPost">
 
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:id="@+id/cl"
