+++ app/src/main/res/layout/update_post.xml
@@ -17,7 +17,7 @@
         android:layout_height="match_parent"
         android:layout_marginTop="?attr/actionBarSize"
         android:layout_marginBottom="?attr/actionBarSize"
-        tools:context=".MainForum.Fragments.FragmentNewPost">
+        tools:context=".MainForum.NewPost.FragmentNewPost">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:id="@+id/cl"
